+++ ICD.Connect.Routing.CrestronPro/DigitalMedia/DmMd64X64/DmMd64X64AdapterSettings.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Connect.Routing.CrestronPro.DigitalMedia.BladeSwitch;
+using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Routing.CrestronPro.DigitalMedia.DmMd64X64
+{
+	[KrangSettings("DmMd64x64", typeof(DmMd64X64Adapter))]
+	public sealed class DmMd64X64AdapterSettings : AbstractCrestronBladeSwitchAdapterSettings
+	{
+	}
+}